Only download stowsh if it doesn't exist

+++ install.sh
@@ -1,10 +1,11 @@
 #!/bin/bash
 cd "$(dirname "${BASH_SOURCE[0]}")" || exit
 
-# get stowsh (and put it in a package)
-curl -o ./bin/stowsh https://raw.githubusercontent.com/williamsmj/stowsh/master/stowsh
+# Get the latest version of stowsh
+[ ! -f ./bin/stowsh ] && curl -o ./bin/stowsh https://raw.githubusercontent.com/williamsmj/stowsh/master/stowsh
 chmod +x ./bin/stowsh
 
+# Packages are non-hidden directories
 pkgs="$(find . -maxdepth 1 ! -name '.*' -type d | sed "s|./||")"
 
 for pkg in $pkgs
